advisor_agent/requirements.txt
# requirements.txt
streamlit
langchain
langchain-groq
langchain_community
langchain-core
langgraph
pydantic
PyYAML
psycopg2-binary
pandas